+++ tests/Fields/InputTypeTest.php
@@ -11,13 +11,14 @@ class InputTypeTest extends FormBuilderTestCase
     {
         $options = [
             'default_value' => 12,
+            'required' => true,
             'help_block' => [
                 'text' => 'this is help'
             ]
         ];
 
         $expectedOptions = $this->getDefaults(
-            [],
+            ['required' => 'required'],
             'hidden_id',
             'Hidden Id',
             12,
@@ -25,6 +26,8 @@ class InputTypeTest extends FormBuilderTestCase
         );
 
         $expectedOptions['help_block']['helpBlockAttrs'] = 'class="help-block" ';
+        $expectedOptions['required'] = true;
+        $expectedOptions['label_attr']['class'] .= ' required';
 
         $expectedViewData = [
             'name' => 'hidden_id',